feat: added put & delete methods
